backend/requirements.txt
# Backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# AI/ML
huggingface-hub==0.20.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tiktoken==0.5.2

# Document Processing
pypdf==3.17.4
python-multipart==0.0.6

# Utilities
httpx==0.26.0
python-dotenv==1.0.0
aiofiles==23.2.1
